:technologist: Add read-only mode to `/api/query/sql`
