+++ kernel/api/sql.go
@@ -44,15 +44,44 @@ func SQL(c *gin.Context) {
 		return
 	}
 
-	var stmt string
-	var skipSingleStatementCheck bool
+	var stmt, mode string
 	if !util.ParseJsonArgs(arg, ret,
 		util.BindJsonArg("stmt", &stmt, true, true),
-		util.BindJsonArg("skipSingleStatementCheck", &skipSingleStatementCheck, false, false),
+		util.BindJsonArg("mode", &mode, false, false),
 	) {
 		return
 	}
-	result, err := sql.Query(stmt, model.Conf.Search.Limit, !skipSingleStatementCheck)
+
+	switch mode {
+	case "":
+		// 默认模式，允许单条语句
+		if err := sql.CheckSingleStatement(stmt); err != nil {
+			ret.Code = -1
+			ret.Msg = err.Error()
+			return
+		}
+	case "readonly":
+		// 只读模式，允许单条语句
+		if err := sql.CheckSingleStatement(stmt); err != nil {
+			ret.Code = -1
+			ret.Msg = err.Error()
+			return
+		}
+		if err := sql.CheckReadonlyStatement(stmt); err != nil {
+			ret.Code = -1
+			ret.Msg = err.Error()
+			return
+		}
+	case "multiple":
+		// 多语句模式，不做校验
+	default:
+		// 未知模式
+		ret.Code = -1
+		ret.Msg = "unknown [mode]"
+		return
+	}
+
+	result, err := sql.Query(stmt, model.Conf.Search.Limit)
 	if err != nil {
 		ret.Code = 1
 		ret.Msg = err.Error()
