+++ kernel/sql/database.go
@@ -1624,7 +1624,10 @@ func SQLTemplateFuncs(templateFuncMap *template.FuncMap) {
 		return
 	}
 	(*templateFuncMap)["querySQL"] = func(stmt string) (ret []map[string]any) {
-		ret, _ = Query(stmt, 1024, true)
+		if err := CheckSingleStatement(stmt); err != nil {
+			return
+		}
+		ret, _ = Query(stmt, 1024)
 		return
 	}
 }
